+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [2.4.0] - TBD
 #### Feature
+- Change venue when match warmup starts
+
+#### Changed
+- Switched from string to time for datetimes to correctly support timezones
 
 ## [2.3.0] - 2022-03-06
 #### Feature